+++ test/functional/ios/ios/mjpeg_server_test.rb
@@ -23,6 +23,18 @@ class AppiumLibCoreTest
         @@driver.update_settings({ mjpegServerScreenshotQuality: -10, mjpegServerFramerate: 60 })
         @@driver.update_settings({ mjpegServerScreenshotQuality: 100, mjpegServerFramerate: 60 })
       end
+
+      def test_start_recording_screen
+        to_path = 'recorded_file_ios.avi'
+        File.delete to_path if File.exist? to_path
+
+        @@driver.start_recording_screen video_type: 'mjpeg'
+        @@driver.find_element(:accessibility_id, 'Buttons').click
+        sleep 2 # second
+        @@driver.back
+        @@driver.stop_and_save_recording_screen to_path
+        assert File.exist? to_path
+      end
     end
   end
 end